requirements.txt
fastapi
loguru
duckdb
pandas
numpy
lancedb
sentence-transformers
pydantic
uvicorn
light-the-torch
typing